requirements.txt
asgiref==3.5.2
Django==4.0.5
pi==0.1.2
Pillow==9.1.1
python-decouple==3.6
qrcode==7.3.1
sqlparse==0.4.2
